+++ CHANGELOG.md
@@ -9,4 +9,5 @@ v2.0.0
 * Automatically choose a light or dark terminal UI color scheme based on the `$COLORFGBG` environment variable. Fallback to no colors. This may be overridden in the config file.
 * Support changing the playback volume with `radio volume` command or a direct API call (the terminal UI or web clients currently do not support changing the volume).
 * Allow to disable figlet banners alltogether, or just the prompt to confirm the choice of the random figlet font for banner (in config file).
+* Fixed sizing issues with ascii-rendered stream art (via pillow), showing stream art can now be deactivated in the config file
 * Switch testing to pytest